credit_risk_system/requirements.txt
# Core AutoGen
pyautogen>=0.2.35,<0.3

# LLM
openai>=1.30.0

# ML
xgboost>=2.0.0
scikit-learn>=1.4.0
shap>=0.44.0
pandas>=2.1.0
numpy>=1.26.0
joblib>=1.4.0
optuna>=3.6.0

# API
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0

# Database
sqlalchemy>=2.0.0

# Config / validation
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.0

# Plotting / evaluation
matplotlib>=3.8.0
seaborn>=0.13.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# CLI
typer>=0.12.0
rich>=13.7.0

# UI
streamlit>=1.35.0
plotly>=5.22.0

# Notebook
jupyter>=1.0.0
ipykernel>=6.29.0
eval_type_backport>=0.2.0
